About project
Project Internet Shop of Sneakers is a web application designed for users to browse and purchase various types of sneakers online. The application is built using the React framework, which is a popular JavaScript library for building user interfaces. The project also uses other web technologies such as JavaScript, HTML, and SCSS to create a visually appealing and user-friendly experience for the users.
The application has a shopping cart feature that allows users to add sneakers to their cart and proceed to checkout. Users can create an account to save their shipping and payment information for future purchases. The checkout process is secure and user-friendly, with various payment options available.

The application also includes a search bar feature that allows users to search for specific sneakers by name or keyword. The search results are displayed in a grid layout, making it easy for users to browse through the available options.

Overall, the Project Internet Shop of Sneakers is a well-designed and functional web application that provides an excellent user experience for sneaker enthusiasts. The use of React, JavaScript, HTML, and SCSS has enabled the creation of a visually appealing and user-friendly application that is sure to attract and retain users.